+++ packages/react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @hrq-ignite-ui/react
 
+## 2.2.0
+
+### Minor Changes
+
+- Component Toast
+
 ## 2.1.0
 
 ### Minor Changes